eval pos intermediate

+++ src/eval/data/evaluationtargetposition.h
@@ -5,12 +5,12 @@ class EvaluationTargetPosition
 {
 public:
     EvaluationTargetPosition() {}
-    EvaluationTargetPosition(float latitude, float longitude, bool has_altitude, float altitude)
+    EvaluationTargetPosition(double latitude, double longitude, bool has_altitude, float altitude)
         : latitude_(latitude), longitude_(longitude), has_altitude_(has_altitude), altitude_(altitude)
     {}
 
-    float latitude_ {0};
-    float longitude_ {0};
+    double latitude_ {0};
+    double longitude_ {0};
     bool has_altitude_ {false};
     float altitude_ {0};
 };
